\name{fpop-package}
\alias{fpop-package}
\alias{fpop}
\docType{package}
\title{Segmentation using optimal partionning and function pruning}
\description{A wrapper to a C implementation of optimal partionning with function pruning}
\details{
\tabular{ll}{Package: \tab fpop\cr
Type: \tab Package\cr
Title: \tab Segmentation using optimal partionning and function pruning\cr
Version: \tab 2014.7.16\cr
Depends: \tab methods, cghseg\cr
SystemRequirements: \tab GNU GSL\cr
Date: \tab 2014-02-26\cr
Author: \tab Guillem Rigaill\cr
Maintainer: \tab Guillem Rigaill <rigaill@evry.inra.fr>\cr
License: \tab LGPL (>= 2.1)\cr}
}
\author{Guillem Rigaill}

\keyword{ package }


